+++ homework/dhrod0325/03/src/components/Recovered/RecoveredTotalList.ts
@@ -1,10 +1,9 @@
-import { calcTotalCountData } from '../lib/utils';
-import { Summary } from '../types';
-import { api } from '../lib/api';
-import { Component } from '../interfaces';
-import { RecoveredTotal } from './RecoveredTotal';
-import { useSpinner } from '../lib/Spinner';
 import { RecoveredList } from './RecoveredList';
+import { Component } from '../../interfaces';
+import { RecoveredTotal } from './RecoveredTotal';
+import { Summary } from '../../types';
+import { useSpinner } from '../../lib/Spinner';
+import { api } from '../../lib/api';
 
 export class RecoveredTotalList implements Component {
   private readonly $total: RecoveredTotal;
